+++ src/directives/ViFocus.js
@@ -1,0 +1,5 @@
+export default {
+    mounted(el) {
+        el.focus()
+    }
+}
