+++ src/directives/index.js
@@ -1,0 +1,8 @@
+import ViFocus from "@/directives/ViFocus";
+import Viintersection from "@/directives/Viintersection";
+
+
+export default [
+    ViFocus,
+    Viintersection
+]
