+++ src/main.js
@@ -2,12 +2,18 @@ import { createApp } from 'vue'
 import App from './App.vue'
 import components from '@/components/UI'
 import router from "@/router/router";
+import directives from "@/directives";
 
 const app = createApp(App)
 
 components.forEach(component => {
     app.component(component.name, component)
 })
+
+directives.forEach(directive => {
+    app.directive(directive.name, directive)
+})
+
 
 app
     .use(router)
