+++ src/pages/PostIdPage.vue
@@ -1,6 +1,6 @@
 <template>
   <div>
-    <h1>This Posts Page with ID = </h1>
+    <h1>This Posts Page with ID = {{ $route.params.id }}}</h1>
   </div>
 </template>
