+++ src/pages/UserPage.vue
@@ -29,9 +29,8 @@
         v-if="!isPostsLoading"
     />
     <div v-else style="padding-left: 200px; color: red">LOADING...</div>
-    <div ref="observer" class="observer">
+    <div v-intersection="loadMorePosts" class="observer"></div>
 
-    </div>
     <!--          <div class="page__wrapper">-->
     <!--            <div-->
     <!--                v-for="pageNumber in totalPages"-->
@@ -133,18 +132,18 @@ export default {
   },
   mounted() {
     this.fetchPosts()
-    console.log(this.$refs.observer)
-    const options = {
-      rootMargin: '0px',
-      threshold: 1.0
-    }
-    const callback = (entries, observer) => {
-      if(entries[0].isIntersecting && this.page < this.totalPages){
-        this.loadMorePosts()
-      }
-    };
-    const observer = new IntersectionObserver(callback, options);
-    observer.observe(this.$refs.observer)
+
+    // const options = {
+    //   rootMargin: '0px',
+    //   threshold: 1.0
+    // }
+    // const callback = (entries, observer) => {
+    //   if(entries[0].isIntersecting && this.page < this.totalPages){
+    //     this.loadMorePosts()
+    //   }
+    // };
+    // const observer = new IntersectionObserver(callback, options);
+    // observer.observe(this.$refs.observer)
   },
   computed: {
     sortedPosts() {
